// packages/sdk/src/views/models/timelineTypes.ts
export enum EventStatus {
    /** The event was not sent and will no longer be retried. */
    NOT_SENT = 'not_sent',
    /** The message is being encrypted */
    ENCRYPTING = 'encrypting',
    /** The event is in the process of being sent. */
    SENDING = 'sending',
    /** The event is in a queue waiting to be sent. */
    QUEUED = 'queued',
    /** The event has been sent to the server, but we have not yet received the echo. */
    SENT = 'sent',
    /** The event was cancelled before it was successfully sent. */
    CANCELLED = 'cancelled',
    /** We received this event */
    RECEIVED = 'received',
}

export interface TimelineEvent {
    eventId: string
    localEventId?: string // if this event was created locally and appended before addEvent, this will be set
    eventNum: bigint
    latestEventId: string // if a message was edited or deleted, this will be set to the latest event id
    latestEventNum: bigint // if a message was edited or deleted, this will be set to the latest event id
    status: EventStatus
    createdAtEpochMs: number // created at times are generated client side, do not trust them
    updatedAtEpochMs?: number // updated at times are generated client side, do not trust them
    content: TimelineEvent_OneOf | undefined // TODO: would be great to have this non optional
    fallbackContent: string
    isEncrypting: boolean // local only, isLocalPending should also be true
    isLocalPending: boolean /// true if we're waiting for the event to get sent back from the server
    isSendFailed: boolean
    confirmedEventNum?: bigint
    confirmedInBlockNum?: bigint
    confirmedAtEpochMs?: number // time miniblock containing this event was created, should be around 2 seconds after the event was created
    threadParentId?: string
    replyParentId?: string
    reactionParentId?: string
    isMentioned: boolean
    isRedacted: boolean
    sender: {
        id: string
    }
    sessionId?: string
}

/// a timeline event should have one or none of the following fields set
export type TimelineEvent_OneOf =
    | ChannelCreateEvent
    | ChannelMessageEncryptedEvent
    | ChannelMessageEncryptedRefEvent
    | ChannelMessageEvent
    | ChannelMessageMissingEvent
    | ChannelPropertiesEvent
    | EncryptedChannelPropertiesEvent
    | FulfillmentEvent
    | InceptionEvent
    | InteractionRequestEvent
    | InteractionRequestEncryptedEvent
    | InteractionResponseEvent
    | KeySolicitationEvent
    | MiniblockHeaderEvent
    | MemberBlockchainTransactionEvent
    | PinEvent
    | ReactionEvent
    | RedactedEvent
    | RedactionActionEvent
    | StreamEncryptionAlgorithmEvent
    | StreamMembershipEvent
    | SpaceImageEvent
    | SpaceUpdateAutojoinEvent
    | SpaceUpdateHideUserJoinLeavesEvent
    | TipEvent
    | TokenTransferEvent
    | SpaceReviewEvent
    | UserBlockchainTransactionEvent
    | UserReceivedBlockchainTransactionEvent
    | UnpinEvent
    | ConversationSeedEvent
    | ConversationSeedResponseEvent
    | SetAppAddressEvent
    | UnsetAppAddressEvent
// ...
export type MentionResult = {
    type: 'mention'
    unread: boolean
    channelId: string
    timestamp: number
    event: TimelineEvent
    thread?: TimelineEvent
}